src/requirements.txt
certifi==2020.6.20
chardet==3.0.4
colored==1.4.2
fritzconnection==1.9.1
idna==2.10
paho-mqtt==1.5.0
requests==2.24.0
urllib3==1.25.10